msm8266-common: Vendor files should be included if-exists
It is standard behavior in CM to only include vendor files 'if-exist'
and therefore, for those possibly only wanting to build a kernel
this change should be used. By not using this, you have to download
vendor files to build even a boot.img. For someone who does not own
the device, this is considered a breach in the blobs agreement and
this person could possibly be held liable for a license violation.

Change-Id: Ic5e0c40b357e895a53a284a5fadd58d131c67113
Signed-off-by: Simon Sickle <simon@simonsickle.com>
Signed-off-by: Louis Popi <theh2o64@gmail.com>